requirements.txt
langgraph==0.6.7
langchain-community==0.3.29
langchain-openai==0.3.33
openai==1.108.0
tavily-python==0.7.12
pandas==2.3.2
sentence-transformers==5.1.0
faiss-cpu==1.12.0
socksio==1.0.0
langchain-google-genai==2.1.12
streamlit==1.50.0